# api/docs/swagger.yaml
definitions:
...
  CreateServiceUserQuery:
    properties:
      email:
        example: joe@gmail.com
        maxLength: 255
        minLength: 2
        type: string
      first_name:
        example: Joe
        maxLength: 255
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 255
        minLength: 2
        type: string
      organization:
        example: ACME
        maxLength: 255
        minLength: 2
        type: string
      password:
        example: password123
        type: string
      role:
        enum:
        - admin
        - user
        - approver
        type: string
    required:
    - email
    - first_name
    - last_name
    - organization
    - password
    - role
    type: object
  CreateUserQuery:
    properties:
      email:
        example: joe@gmail.com
        maxLength: 255
        minLength: 2
        type: string
      first_name:
        example: Joe
        maxLength: 255
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 255
        minLength: 2
        type: string
      organization:
        example: ACME
        maxLength: 255
        minLength: 2
        type: string
      role:
        enum:
        - admin
        - user
        - approver
        type: string
    required:
    - email
    - first_name
    - last_name
    - organization
    - role
    type: object
  ErrorResponse:
    properties:
      error:
        description: Error message
        example: Some error message
        type: string
    type: object
  JSendError:
    properties:
      message:
        description: Error message
        example: Internal server error
        type: string
      status:
        description: Status
        example: error
        type: string
    type: object
  JSendFailure-ErrorResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/ErrorResponse'
        description: Data with error message(s)
      status:
        description: Status 'fail'
        example: fail
        type: string
    type: object
  JSendFailure-ValidationResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/ValidationResponse'
        description: Data with error message(s)
      status:
        description: Status 'fail'
        example: fail
        type: string
    type: object
...
  JSendSuccess-map_string_string:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/map_string_string'
        description: Data with success message(s)
      status:
        description: Status 'success'
        example: success
        type: string
    type: object
...
  ValidationResponse:
    properties:
      errors:
        description: Validation errors
        items:
          $ref: '#/definitions/apierr.ValidationError'
        type: array
    type: object
  apierr.ValidationError:
    properties:
      field:
        type: string
      reason:
        type: string
    type: object
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
paths:
  /admin/users:
    post:
      description: |-
        __Admin role required__
        Create a new user for the API. Ths user will receive an email with a token to set their password.
        You can create users with the following roles: `admin`, `user`, `approver`.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateUserQuery'
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: Non-admin user
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      security:
      - Bearer: []
      summary: Create a new user
      tags:
      - Admin
  /admin/users/service:
    post:
      description: |-
        __Admin role required__
        Create a new service user for the API.
        You can create users with the following roles: `admin`, `user`, `approver`.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateServiceUserQuery'
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: Non-admin user
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      security:
      - Bearer: []
      summary: Create a new service user
      tags:
      - Admin
...
swagger: "2.0"